PYIC-5506: Refactor - move CriApiService to libs
