+++ .secrets.baseline
@@ -427,15 +427,6 @@
         "line_number": 1590
       }
     ],
-    "lambdas/process-cri-callback/src/main/java/uk/gov/di/ipv/core/processcricallback/service/CriApiService.java": [
-      {
-        "type": "Secret Keyword",
-        "filename": "lambdas/process-cri-callback/src/main/java/uk/gov/di/ipv/core/processcricallback/service/CriApiService.java",
-        "hashed_secret": "b894b81be94cf8fa8d7536475aaec876addf05c8",
-        "is_verified": false,
-        "line_number": 52
-      }
-    ],
     "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/pact/addressCri/ContractTest.java": [
       {
         "type": "JSON Web Token",
@@ -1486,36 +1477,6 @@
         "line_number": 268
       }
     ],
-    "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/service/CriApiServiceTest.java": [
-      {
-        "type": "Secret Keyword",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/service/CriApiServiceTest.java",
-        "hashed_secret": "b894b81be94cf8fa8d7536475aaec876addf05c8",
-        "is_verified": false,
-        "line_number": 65
-      },
-      {
-        "type": "Secret Keyword",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/service/CriApiServiceTest.java",
-        "hashed_secret": "767ef7376d44bb6e52b390ddcd12c1cb1b3902a4",
-        "is_verified": false,
-        "line_number": 66
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/service/CriApiServiceTest.java",
-        "hashed_secret": "29cf87d79251ad6757dbaa452e459c733ca0c819",
-        "is_verified": false,
-        "line_number": 68
-      },
-      {
-        "type": "Base64 High Entropy String",
-        "filename": "lambdas/process-cri-callback/src/test/java/uk/gov/di/ipv/core/processcricallback/service/CriApiServiceTest.java",
-        "hashed_secret": "bbd2d902a5f8ca57be23f5280ec809e60b1275f2",
-        "is_verified": false,
-        "line_number": 196
-      }
-    ],
     "libs/common-services/src/main/java/uk/gov/di/ipv/core/library/service/ConfigService.java": [
       {
         "type": "Secret Keyword",
@@ -1704,6 +1665,45 @@
         "hashed_secret": "d1d4766fd2a2f73940b46f0d7ccabba3dd98e2cd",
         "is_verified": false,
         "line_number": 68
+      }
+    ],
+    "libs/cri-api-service/src/main/java/uk/gov/di/ipv/core/library/criapiservice/CriApiService.java": [
+      {
+        "type": "Secret Keyword",
+        "filename": "libs/cri-api-service/src/main/java/uk/gov/di/ipv/core/library/criapiservice/CriApiService.java",
+        "hashed_secret": "b894b81be94cf8fa8d7536475aaec876addf05c8",
+        "is_verified": false,
+        "line_number": 51
+      }
+    ],
+    "libs/cri-api-service/src/test/java/uk/gov/di/ipv/core/library/criapiservice/CriApiServiceTest.java": [
+      {
+        "type": "Secret Keyword",
+        "filename": "libs/cri-api-service/src/test/java/uk/gov/di/ipv/core/library/criapiservice/CriApiServiceTest.java",
+        "hashed_secret": "b894b81be94cf8fa8d7536475aaec876addf05c8",
+        "is_verified": false,
+        "line_number": 65
+      },
+      {
+        "type": "Secret Keyword",
+        "filename": "libs/cri-api-service/src/test/java/uk/gov/di/ipv/core/library/criapiservice/CriApiServiceTest.java",
+        "hashed_secret": "767ef7376d44bb6e52b390ddcd12c1cb1b3902a4",
+        "is_verified": false,
+        "line_number": 66
+      },
+      {
+        "type": "Base64 High Entropy String",
+        "filename": "libs/cri-api-service/src/test/java/uk/gov/di/ipv/core/library/criapiservice/CriApiServiceTest.java",
+        "hashed_secret": "29cf87d79251ad6757dbaa452e459c733ca0c819",
+        "is_verified": false,
+        "line_number": 68
+      },
+      {
+        "type": "Base64 High Entropy String",
+        "filename": "libs/cri-api-service/src/test/java/uk/gov/di/ipv/core/library/criapiservice/CriApiServiceTest.java",
+        "hashed_secret": "bbd2d902a5f8ca57be23f5280ec809e60b1275f2",
+        "is_verified": false,
+        "line_number": 196
       }
     ],
     "libs/evcs-service/src/main/java/uk/gov/di/ipv/core/library/client/EvcsClient.java": [
