+++ lambdas/process-cri-callback/build.gradle
@@ -15,6 +15,7 @@ dependencies {
 			project(":libs:audit-service"),
 			project(":libs:cimit-service"),
 			project(":libs:common-services"),
+			project(":libs:cri-api-service"),
 			project(":libs:cri-response-service"),
 			project(":libs:cri-storing-service"),
 			project(":libs:gpg45-evaluator"),
