+++ libs/cri-api-service/src/main/java/uk/gov/di/ipv/core/library/criapiservice/exception/CriApiException.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.core.processcricallback.exception;
+package uk.gov.di.ipv.core.library.criapiservice.exception;
 
 import lombok.Getter;
 import uk.gov.di.ipv.core.library.domain.ErrorResponse;
