+++ settings.gradle
@@ -22,6 +22,7 @@ include "lambdas",
 		"libs:audit-service",
 		"libs:cimit-service",
 		"libs:common-services",
+		"libs:cri-api-service",
 		"libs:cri-response-service",
 		"libs:cri-storing-service",
 		"libs:email-service",
